[update] move unit test in decode/inst_16.rs to c_extension.rs
